fix: update state & add utils
